add admin restrictions to edit and delete article

+++ view/frontend/articleView.php
@@ -4,8 +4,10 @@
 <h1>Mon super blog !</h1>
 <p><a href="index.php">Retour à la liste des billets</a></p>
 
+<?php if($user->getStatus() === 'admin') { ?>
 <p><a href="index.php?action=editArticle&amp;id=<?= $article->getId() ?>">Modifier l'article</a></p>
 <p><a href="index.php?action=deleteArticle&amp;id=<?= $article->getId() ?>">Supprimer l'article</a></p>
+<?php } ?>
 
 <div class="news">
     <h3>
@@ -37,9 +39,10 @@ foreach ($comments as $data)
 ?>
     <p><strong><?= htmlspecialchars($data->getUserId()) ?></strong> le <?= $data->getCreationDate(); ?></p>
     <p><?= nl2br(htmlspecialchars($data->getContent())) ?></p>
+<?php if($user->getStatus() === 'admin') { ?>
     <p><a href="index.php?action=deleteComment&amp;id=<?= $data->getId()?>&amp;article=<?= $article->getId()?>">Supprimer le commentaire</a></p>
 <?php
-}
+}}
 ?>
 <?php $content = ob_get_clean(); ?>
